Worked on leds some more

+++ Visualizer/Led.h
@@ -6,7 +6,7 @@ public:
 	Led();
 	sf::Color getColor();
 	~Led();
-	int id;
+	int8_t id;
 	int red = 0;
 	int green = 0;
 	int blue = 0;
